backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
gtfs-realtime-bindings==1.0.0
typing-inspection==0.4.1
typing_extensions==4.13.2
SQLAlchemy==2.0.41
sqlmodel==0.0.24